requirements.txt
Django>=4.2,<5.0
djangorestframework
djangorestframework-simplejwt
drf-spectacular>=0.27.0
psycopg2-binary
Pillow
django-filter
python-dotenv
requests>=2.31.0
uvicorn>=0.30.0